Refactor DDGLogoListener to just hide-show background image
