+++ app/src/main/res/layout/include_new_browser_tab.xml
@@ -21,7 +21,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:layout_marginTop="?attr/actionBarSize"
-    android:animateLayoutChanges="true"
     android:foreground="@android:color/transparent"
     tools:context="com.duckduckgo.app.browser.BrowserActivity"
     tools:showIn="@layout/fragment_browser_tab">
@@ -29,15 +28,14 @@
     <ImageView
         android:id="@+id/ddgLogo"
         android:layout_width="0dp"
-        android:alpha="0"
         android:layout_height="wrap_content"
         android:contentDescription="@string/duckDuckGoLogoDescription"
         android:src="@drawable/logo_full"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintBottom_toTopOf="@+id/ctaContainer"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHeight_max="180dp"
         app:layout_constraintStart_toStartOf="parent"
+        android:layout_marginTop="100dp"
+        android:alpha="0"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintWidth_max="180dp"
         android:visibility="gone"/>
